Move files to their respective modules
We are doing this have a more modulated structure, from now on when we want to make certain changes we'll make them in their respective modules.
