+++ src/ui/utils.js
@@ -1,50 +1,29 @@
-import { currenciesName } from "../utils/currency-name.js";
-export function createExchangeTable(rates) {
-  const $exchangeTable = document.querySelector("#exchange-table-body");
-  const $rates = Object.keys(rates);
+import { handleInputCurrency, handleInputDate, handleListChange } from "./handlers.js";
+import { displayExchangeUI } from "./exchange.js";
 
-  $rates.forEach((rate, index) => {
-    const $row = document.createElement("tr");
-    const $currencyNumber = document.createElement("th");
-    const $currencyCode = document.createElement("th");
-    const $currencyFullName = document.createElement("th");
-    const $rate = document.createElement("th");
-
-    $currencyNumber.innerText = `${index + 1}`;
-    $currencyCode.innerText = `${rate}`;
-    $currencyFullName.innerText = `${currenciesName[index]}`;
-    $rate.innerText = `$${rates[rate]}`;
-    $row.appendChild($currencyNumber);
-    $row.appendChild($currencyCode);
-    $row.appendChild($currencyFullName);
-    $row.appendChild($rate);
-    $exchangeTable.appendChild($row);
-  });
+export function setupListChanges(list, date) {
+  list.onclick = (currency) => {
+    const clickedCurrency = handleListChange(currency);
+    const selectedDate = handleInputDate(date);
+    displayExchangeUI(clickedCurrency, selectedDate);
+  };
 }
 
-export function createCurrencyList(rates) {
-  const $currencyList = document.querySelector("#currency-list");
-  const currencies = Object.keys(rates);
-
-  currencies.forEach((currency, index) => {
-    const $list = document.createElement("li");
-    const $item = document.createElement("a");
-    $item.className = "dropdown-item text-start";
-    $item.href = "#";
-    $item.innerText = `${currency} - ${currenciesName[index]}`;
-
-    $list.appendChild($item);
-    $currencyList.appendChild($list);
-  });
+export function setupCurrencyChanges(currency) {
+  currency.oninput = () => {
+    const selectedCurrency = handleInputCurrency(currency);
+    displayExchangeUI(selectedCurrency);
+  };
 }
 
-export function clearExchangeTable() {
-  document.querySelector("#exchange-table-body").innerHTML = "";
-}
-
-export function getCurrencyCode(currency) {
-  const currencyCode = currency.innerText.substring(0, 3);
-  return currencyCode;
+export function setupDateChanges(date, currency) {
+  date.oninput = () => {
+    setTimeout(() => {
+      const selectedCurrency = handleInputCurrency(currency);
+      const selectedDate = handleInputDate(date);
+      displayExchangeUI(selectedCurrency, selectedDate);
+    }, 1000);
+  };
 }
 
 export function setCurrencyTitle(base, currency) {
@@ -65,4 +44,9 @@ export function setDateTitle(present, date) {
 
   if (currentInputDate && currentInputDate !== "latest") return (currentDateTitle.innerText = `At ${currentInputDate} as date of exchange`);
   return (currentDateTitle.innerText = `At ${presentDate} as date of exchange`);
+}
+
+export function getCurrencyCode(currency) {
+  const currencyCode = currency.innerText.substring(0, 3);
+  return currencyCode;
 }
